New version of findPath.py
This version loads into memory only a part of the galaxy then compiles
the list of neighbors for each system. This is a faster method of
finding whether a path exists and I feel stupid for not implementing
this logic when I came up with it (I thought it would take way more
memory so I didn't even bother implementing it).